+++ src/proposal.rs
@@ -1,11 +1,13 @@
-use crate::{block::N_BLOCKS, helper::UpgradeHelper, inputs::get_time_string, network::Network};
-use handlebars::{Handlebars, RenderError};
+use crate::{
+    block::N_BLOCKS, errors::ProposalError, helper::UpgradeHelper, inputs::get_time_string,
+    network::Network,
+};
+use handlebars::Handlebars;
 use num_format::ToFormattedString;
 use serde_json::json;
 
-/// Prepares the proposal text by filling in the necessary information
-/// to the proposal template.
-pub fn prepare_proposal(helper: &UpgradeHelper) -> Result<String, RenderError> {
+/// Renders the proposal template, filling in the necessary information.
+pub fn render_proposal(helper: &UpgradeHelper) -> Result<String, ProposalError> {
     let mut handlebars = Handlebars::new();
     handlebars.set_strict_mode(true);
 
@@ -33,15 +35,7 @@ pub fn prepare_proposal(helper: &UpgradeHelper) -> Result<String, RenderError> {
         "voting_time": helper.voting_period,
     });
 
-    handlebars.render("proposal", &data)
-}
-
-/// Writes the proposal contents to a file.
-pub fn write_proposal_to_file(
-    proposal: &str,
-    proposal_file_name: &str,
-) -> Result<(), std::io::Error> {
-    std::fs::write(proposal_file_name, proposal)
+    Ok(handlebars.render("proposal", &data)?)
 }
 
 /// Returns the appropriate Markdown link to the block on Mintscan for the given network and height.
@@ -77,34 +71,14 @@ mod tests {
     use chrono::Utc;
 
     #[test]
-    fn test_prepare_proposal_pass() {
+    fn test_render_proposal_pass() {
         let helper = UpgradeHelper::new(Network::Mainnet, "v0.0.1", "v0.1.0", Utc::now(), 60);
 
-        let result = prepare_proposal(&helper);
+        let result = render_proposal(&helper);
         assert!(
             result.is_ok(),
             "Error rendering proposal: {}",
             result.unwrap_err(),
         );
-    }
-
-    #[test]
-    fn test_write_proposal_to_file_pass() {
-        let proposal_file_name = format!("proposal-{}-{}.md", Network::Mainnet, "v0.1.0");
-        let result = write_proposal_to_file("test", proposal_file_name.as_str());
-        assert!(
-            result.is_ok(),
-            "Error writing proposal to file: {}",
-            result.unwrap_err(),
-        );
-
-        // Check that file exists
-        assert!(
-            std::path::Path::new(proposal_file_name.as_str()).exists(),
-            "Proposal file does not exist",
-        );
-
-        // Clean up
-        std::fs::remove_file(proposal_file_name).unwrap();
     }
 }