+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 
 ### Features
 
+- (cli) [#1](https://github.com/evmos/proposer/pull/1) Prepare CLI command to submit proposal.
+
+### Improvements
+
+- (cli) [#12](https://github.com/evmos/proposer/pull/12) Use [clap](https://github.com/clap-rs/clap) for CLI handling.
+- (all) [#13](https://github.com/evmos/proposer/pull/13) Implement better error handling using [thiserror](https://github.com/dtolnay/thiserror).
 - (tests) [#11](https://github.com/evmos/proposer/pull/11) Mock API responses for tests.
-- (cmd) [#1](https://github.com/evmos/proposer/pull/1) Prepare CLI command to submit proposal.
 
 ## Legacy Changelog from [Original Repo](https://github.com/MalteHerrmann/upgrade-helper)